GaitSimulation/requirements.txt
appnope==0.1.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
bleach==3.0.2
configparser==3.5.0
decorator==4.3.0
defusedxml==0.5.0
entrypoints==0.2.3
enum34==1.1.6
functools32==3.2.3.post2
futures==3.2.0
ipaddress==1.0.22
ipykernel==4.10.0
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.4
jupyter-core==4.4.0
MarkupSafe==1.1.0
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.4
numpy==1.15.4
open3d-python==0.4.0.0
pandocfilters==1.4.2
pathlib2==2.3.3
pexpect==4.6.0
pickleshare==0.7.5
prometheus-client==0.5.0
prompt-toolkit==1.0.15
ptyprocess==0.6.0
Pygments==2.3.1
PyODE==1.2.1
python-dateutil==2.7.5
pyzmq==17.1.2
scandir==1.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
